+++ packages/core/test/render3/perf/directive_inputs/index.ts
@@ -6,7 +6,7 @@
  * found in the LICENSE file at https://angular.io/license
  */
 import {ɵɵadvance, ɵɵdefineDirective, ɵɵelement, ɵɵproperty} from '../../../../src/render3/index';
-import {refreshView} from '../../../../src/render3/instructions/shared';
+import {refreshView} from '../../../../src/render3/instructions/change_detection';
 import {RenderFlags} from '../../../../src/render3/interfaces/definition';
 import {TVIEW} from '../../../../src/render3/interfaces/view';
 import {createBenchmark} from '../micro_bench';
