+++ packages/core/test/render3/perf/map_based_style_and_class_bindings/index.ts
@@ -5,9 +5,10 @@
  * Use of this source code is governed by an MIT-style license that can be
  * found in the LICENSE file at https://angular.io/license
  */
+
 import {ɵɵadvance} from '../../../../src/render3/instructions/advance';
+import {refreshView} from '../../../../src/render3/instructions/change_detection';
 import {ɵɵelement, ɵɵelementEnd, ɵɵelementStart} from '../../../../src/render3/instructions/element';
-import {refreshView} from '../../../../src/render3/instructions/shared';
 import {ɵɵclassMap, ɵɵstyleMap} from '../../../../src/render3/instructions/styling';
 import {RenderFlags} from '../../../../src/render3/interfaces/definition';
 import {TVIEW} from '../../../../src/render3/interfaces/view';
