+++ packages/core/test/render3/perf/noop_change_detection/index.ts
@@ -5,7 +5,8 @@
  * Use of this source code is governed by an MIT-style license that can be
  * found in the LICENSE file at https://angular.io/license
  */
-import {refreshView} from '../../../../src/render3/instructions/shared';
+
+import {refreshView} from '../../../../src/render3/instructions/change_detection';
 import {TVIEW} from '../../../../src/render3/interfaces/view';
 import {createBenchmark} from '../micro_bench';
 import {setupRootViewWithEmbeddedViews} from '../setup';
