+++ packages/core/test/render3/view_fixture.ts
@@ -11,7 +11,9 @@ import {EffectManager} from '@angular/core/src/render3/reactivity/effect';
 import {stringifyElement} from '@angular/platform-browser/testing/src/browser_util';
 
 import {extractDirectiveDef} from '../../src/render3/definition';
-import {createLView, createTNode, createTView, refreshView, renderView} from '../../src/render3/instructions/shared';
+import {refreshView} from '../../src/render3/instructions/change_detection';
+import {renderView} from '../../src/render3/instructions/render';
+import {createLView, createTNode, createTView} from '../../src/render3/instructions/shared';
 import {DirectiveDef, DirectiveDefList, DirectiveTypesOrFactory, PipeDef, PipeDefList, PipeTypesOrFactory, RenderFlags} from '../../src/render3/interfaces/definition';
 import {TConstants, TElementNode, TNodeType} from '../../src/render3/interfaces/node';
 import {LView, LViewFlags, TView, TViewType} from '../../src/render3/interfaces/view';
